backend_django/test_login_api.py
import requests
import json

url = "http://localhost:8000/api/admin/login"
data = {
    "username": "admin",
    "password": "Admin@456"
}

try:
    response = requests.post(url, json=data)
    print(f"Status Code: {response.status_code}")
    print(f"Response: {json.dumps(response.json(), indent=2)}")
except Exception as e:
    print(f"Error: {e}")
    print(f"Response text: {response.text if 'response' in locals() else 'No response'}")
